Correct Range with Mountains

+++ FE_v1/Infantry.cpp
@@ -5,6 +5,6 @@ Infantry::Infantry()
 	
 }
 
-int Infantry::costeMovimiento(Tile* tile) {
+float Infantry::costeMovimiento(Tile* tile) {
 	return tile->movementCost;
 }
