+++ FE_v1/Infantry.h
@@ -6,5 +6,5 @@
 class Infantry : public MovementType {
 public:
 	Infantry();
-	int costeMovimiento(Tile* tile) override;
+	float costeMovimiento(Tile* tile) override;
 };
