+++ FE_v1/MovementType.cpp
@@ -4,7 +4,7 @@ MovementType::MovementType(int movementRange) {
 	this->movementRange = movementRange;
 }
 
-int MovementType::costeMovimiento(Tile* tile) {
+float MovementType::costeMovimiento(Tile* tile) {
 	cout << "que cojones" << endl;
 	return 0;
 }
