+++ FE_v1/MovementType.h
@@ -6,7 +6,7 @@ class MovementType {
 public:
 	MovementType(int movementRange);
 
-	virtual int costeMovimiento(Tile* tile);
+	virtual float costeMovimiento(Tile* tile);
 
 	int movementRange;
 };
